update npm packages
- TS to 3.9
- Next to 9.4
    - required adding undefined to possible types of router.query.id
- eslint-plugin-react-hooks
    - added to eslint config, and fixed errors it found
    - refactored Resize effect into component (so it has Redux access)
- typing updates
