+++ packages/frontend/package.json
@@ -28,7 +28,7 @@
     "@testing-library/react": "^9.4.0",
     "@types/jest": "^25.2.1",
     "@types/next-redux-saga": "^3.0.2",
-    "@types/react-redux": "^7.1.7",
+    "@types/react-redux": "^7.1.8",
     "@zeit/next-css": "^1.0.1",
     "axios": "^0.19.2",
     "class-transformer": "^0.2.3",
@@ -38,7 +38,7 @@
     "isomorphic-unfetch": "^3.0.0",
     "jest": "^25.1.0",
     "js-cookie": "^2.2.1",
-    "next": "^9.3.4",
+    "next": "^9.4.0",
     "next-redux-saga": "^4.1.2",
     "next-redux-wrapper": "^5.0.0",
     "protobufjs": "~6.8.9",
@@ -52,15 +52,15 @@
     "semantic-ui-react": "^0.88.2",
     "socket.io-client": "^2.3.0",
     "styled-jsx": "^3.2.5",
-    "sweetalert2": "^9.10.10",
-    "ts-jest": "^25.2.0"
+    "sweetalert2": "^9.10.13",
+    "ts-jest": "^25.5.1"
   },
   "devDependencies": {
     "@babel/plugin-proposal-decorators": "^7.8.3",
     "@types/js-cookie": "^2.2.6",
-    "@types/node": "^13.7.0",
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
+    "@types/node": "^13.13.6",
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
     "@types/socket.io-client": "^1.4.32",
     "@types/styled-jsx": "^2.2.8",
     "@typescript-eslint/eslint-plugin": "^2.19.0",
@@ -76,12 +76,12 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^1.7.0",
+    "eslint-plugin-react-hooks": "^4.0.2",
     "file-loader": "^5.0.2",
     "lint-staged": "^10.0.7",
     "prettier": "^1.19.1",
     "redux-saga-test-plan": "^4.0.0-rc.3",
-    "typescript": "^3.7.5",
+    "typescript": "^3.9.2",
     "url-loader": "^3.0.0"
   }
 }
